{"type":"TI", "runs":"2"}

+++ exerciseCh5.java
@@ -14,12 +14,12 @@ public class exerciseCh5 extends ConsoleProgram {
 	
 	private void exerciseTwo() {
 		println ("Enter coeficients for this quadratic equation: ");
-		double a = readDouble ("a: ");
-		double b = readDouble ("b: ");
-		double c = readDouble ("c: ");
-		double condition = Math.sqrt((b*b)-(4*a*c));
-		double firstSolution;
-		double secondSolution;
+		int a = readInt ("a: ");
+		int b = readInt ("b: ");
+		int c = readInt ("c: ");
+		int condition = Math.sqrt((b*b)-(4*a*c));
+		int firstSolution;
+		int secondSolution;
 		if (condition>0) {
 			firstSolution = (-b + condition) / (2*a);
 			secondSolution = (-b - condition) / (2*a);
